sam: extract the XRAS message builder out of the blueprint
## Summary

New `src/sam/queries/xras_notices.py`, mirroring
`sam/queries/expiration_notices.py`. Moves `_XRAS_KIND_SUBJECTS`,
`_load_xras_action`, `_action_increments` and `_xras_messages` out of
`webapp/dashboards/allocations/blueprint.py`, where they were private
closures reachable only through a Flask request context.

That location was the blocker for automatic sending, not the shape: the
XRAS pre-deploy smoke correctly noted the builders "already take an
action and are shaped for a handler to call", but `src/scheduling/` is
AST-gated against importing Flask, so nothing under it could call them.

Only two kinds of coupling had to go, and both had a worked precedent in
`build_expiration_messages`:

- `db.session` -> a `session` parameter (three sites);
- `current_user.username` -> `requested_by=`, keyword-only and
  **undefaulted** — a default in an audit column is a plausible-looking
  lie.

`xras_notify_form` and `xras_notify` now go through the same function, so
the button and the coming task cannot drift on the dedup key. That key is
the entire safety argument for running the two side by side: identical
keys mean the ledger suppresses whichever fires second, so no locking or
claiming is needed around the card.

## Not exported from `sam/queries/__init__.py`

Deliberately, and the module docstring says so: that file imports its
submodules eagerly, and this one imports `sam.notify` at module scope, so
listing it would put `sam.notify.base` into the graph of every
`from sam.queries import ...` in the tree. The blueprint imports it by
full dotted path.

The trap worth naming is that `xras_activation.py` **is** exported, and
safely — it does not import `sam.notify`. The two modules look alike and
have to be treated differently.

## One test moved, and why it is not a shim

`test_xras_notify.py` (the routes, 28 tests) passes **unedited** — that is
the proof the move was pure. `test_xras_dashboard.py` needed one edit:
its `TestSignedIncrements` class reached past the routes into
`blueprint._action_increments` directly, so it moved to the new
`test_xras_notices_builder.py` with the helper it exercises. The tests got
*shorter* in the move — they no longer need `app.app_context()`, which is
the visible half of what the extraction bought. Six cases added while it
was open, covering the unparseable-payload paths that yield `[]` rather
than a guess.

## Test Results

`test_xras_notify.py`, `test_xras_dashboard.py`,
`test_xras_notices_builder.py`, `test_notify_import_graph.py` —
146 passed, 1 skipped.

Import probe with `FLASK_ACTIVE` stripped: importing
`sam.queries.xras_notices` adds none of click / flask / rich / kubernetes
/ webapp / cli.
