+++ src/webapp/dashboards/allocations/blueprint.py
@@ -12,7 +12,6 @@ from flask import (
 )
 from flask_login import login_required, current_user
 from datetime import datetime, timedelta
-import json
 from typing import List, Dict
 
 from webapp.extensions import db, cache, user_aware_cache_key
@@ -23,11 +22,7 @@ from webapp.utils.htmx import (
     read_sort, read_theme, register_typeahead,
 )
 from webapp.api.xras.recheck import recheck_action
-from sam import fmt
-from sam.enums import ResourceTypeName
-from sam.integration.xras import (
-    XrasActionLog, XrasActivationEvent, XrasResourceRepositoryKeyResource,
-)
+from sam.integration.xras import XrasActivationEvent
 from sam.manage.transaction import management_transaction
 from sam.projects.projects import Project
 from webapp.utils.notify import get_notifier, notify_summary
@@ -57,13 +52,15 @@ from sam.queries.xras_actions import (
 )
 from sam.queries.xras_activation import (
     ACTIVITY_TAGS,
-    XRAS_SERVICE_KINDS,
     get_latest_xras_action_id,
     get_xras_activation_events,
     get_xras_activity,
     get_xras_pending_recipients,
-    xras_dedup_key,
 )
+# Full dotted path, never through `sam.queries` — that package imports its
+# submodules eagerly, and this one imports `sam.notify`. See the module
+# docstring; `tests/unit/test_notify_import_graph.py` is the gate.
+from sam.queries.xras_notices import build_xras_messages, load_xras_action
 from sam.queries.usage_cache import cached_allocation_usage, purge_usage_cache, usage_cache_info
 from sam.queries.lookups import find_project_by_code
 from sam.schemas.forms import CreateChargeAdjustmentForm, XrasActivationEventForm
@@ -1658,160 +1655,17 @@ def _record_activation_event(project, event_type, *, comment=None,
     )
 
 
-#: kind → (subject template, headline verb for the template context).
-#: The subject lives here rather than in the Jinja file because it is also the
-#: `notification_log.subject` column an operator reads back in the admin log,
-#: and a subject assembled inside a template cannot be searched from SQL.
-_XRAS_KIND_SUBJECTS = {
-    'xras_activation': 'NSF NCAR Project {projcode} is now active',
-    'xras_supplement': 'NSF NCAR Project {projcode} has received additional allocation',
-    'xras_extension': 'NSF NCAR Project {projcode} allocation has been extended',
-    'xras_update': 'NSF NCAR Project {projcode} allocation has been renewed',
-    # Deliberately directionless: an Adjustment can subtract, and a subject
-    # line promising good news is read long before the body corrects it.
-    'xras_adjustment': 'NSF NCAR Project {projcode} allocation has been adjusted',
-}
-
-
-def _load_xras_action(action_id):
-    """One ``xras_action_log`` row, or None. No permission logic — callers gate."""
-    if action_id is None:
-        return None
-    return db.session.get(XrasActionLog, action_id)
-
-
-def _action_increments(action, *, signed=False):
-    """What *this* action changed, read back off its own stored payload.
-
-    A supplement's mail has to say how much was added, and that number exists
-    nowhere else: the allocation now holds the **new total**, and
-    ``allocation_transaction`` records the delta without naming the XRAS
-    action. The payload is the only place the increment survives, which is one
-    more reason ``raw_payload`` is stored verbatim.
-
-    ``signed=True`` prefixes a ``+`` on positive amounts, for the Adjustment
-    mail. An Adjustment is the **only** action type whose amounts can be
-    negative (``AdjustmentHandler`` exists to honour them — the legacy
-    factory's copy-pasted ``> 0`` gate is what kept it dark), so this is the
-    one message where the reader cannot infer the direction from the action
-    type and has to be shown it. ``fmt.number`` already carries the minus.
-
-    Units are computed on the **magnitude**: ``allocation_unit`` decides
-    singular/plural from the value, and -1 is one hour in either direction.
-
-    Returns ``[{'resource_name', 'amount', 'units'}]``, or ``[]`` for anything
-    unparseable — a wrong number here would be worse than an absent one, so
-    every failure path yields nothing rather than a guess.
-    """
-    if action is None or not action.raw_payload:
-        return []
-    try:
-        payload = json.loads(action.raw_payload)
-    except (ValueError, TypeError):
-        return []
-
-    wire = payload.get('resources') or []
-    keys = [w.get('resourceRepositoryKey') for w in wire
-            if w.get('resourceRepositoryKey') is not None]
-    if not keys:
-        return []
-
-    mapped = {
-        m.resource_repository_key: m.resource
-        for m in db.session.query(XrasResourceRepositoryKeyResource)
-        .filter(XrasResourceRepositoryKeyResource
-                .resource_repository_key.in_(keys)).all()
-    }
-
-    out = []
-    for item in wire:
-        resource = mapped.get(item.get('resourceRepositoryKey'))
-        if resource is None:
-            continue
-        try:
-            amount = float(item.get('awardedAmount'))
-        except (TypeError, ValueError):
-            continue
-        shown = fmt.number(amount)
-        if signed and amount > 0:
-            shown = f'+{shown}'
-        out.append({
-            'resource_name': resource.resource_name,
-            'amount': shown,
-            'units': ResourceTypeName.allocation_unit(
-                resource.resource_type.resource_type
-                if resource.resource_type else None, abs(amount)),
-        })
-    return sorted(out, key=lambda r: r['resource_name'])
-
-
 def _xras_messages(project, people, *, action=None):
-    """Build one :class:`~sam.notify.Message` per recipient for one XRAS action.
+    """The route's binding of :func:`~sam.queries.xras_notices.build_xras_messages`.
 
-    ``dedup_key`` embeds the action, so a Supplement mints a different key from
-    the New that preceded it: each outcome can be reported once, and re-opening
-    the modal about the same one cannot re-mail anybody. That is the same key
-    the activity table reads back to decide whether a row says "notified".
-
-    ``action=None`` falls back to the newest action naming the project, which
-    is what the Notify button did before it became action-aware and what a
-    caller with only a project id still gets.
+    Two lines, and they are the two the scheduled task cannot supply: Flask's
+    request-scoped session, and the operator who clicked. Everything else —
+    the payload, the subject, and above all the dedup key — lives in
+    ``sam.queries.xras_notices`` so that the button and ``xras_notices`` can
+    never disagree about what has already been sent.
     """
-    from sam.notify import Message, to_recipients
-
-    if action is None:
-        action = _load_xras_action(
-            get_latest_xras_action_id(db.session, project.project_id))
-
-    action_id = action.xras_action_log_id if action is not None else None
-    kind = XRAS_SERVICE_KINDS.get((action.service or '') if action else '',
-                                  'xras_activation')
-
-    usage = project.get_detailed_allocation_usage()
-    resources = [{
-        'resource_name': name,
-        'amount': fmt.number(info.get('allocated')),
-        'units': ResourceTypeName.allocation_unit(info.get('resource_type'),
-                                                  info.get('allocated')),
-        'end_date': fmt.date_str(info.get('end_date'), null=None),
-    } for name, info in sorted(usage.items())]
-
-    lead_email = project.lead.primary_email if project.lead else None
-    context = {
-        'project_code': project.projcode,
-        'project_title': project.title,
-        'project_lead': project.lead.display_name if project.lead else 'Project Lead',
-        'project_lead_email': lead_email,
-        'resources': resources,
-        # Only one template reads each of these, but every kind carries both —
-        # a template that renders an undefined name renders nothing, silently,
-        # so the cheapest guard is for the key to always exist.
-        'added': _action_increments(action) if kind == 'xras_supplement' else [],
-        # Signed, and separate from `added` on purpose: `added` is a promise
-        # that every number in it is an increase, which the supplement wording
-        # leans on. An adjustment makes no such promise.
-        'changes': (_action_increments(action, signed=True)
-                    if kind == 'xras_adjustment' else []),
-        'action_type': action.action_type if action else None,
-    }
-    subject = _XRAS_KIND_SUBJECTS.get(
-        kind, _XRAS_KIND_SUBJECTS['xras_activation']
-    ).format(projcode=project.projcode)
-
-    return [
-        Message(
-            kind=kind,
-            recipient=recipient,
-            subject=subject,
-            context=context,
-            entity=('project', project.project_id),
-            projcode=project.projcode,
-            dedup_key=xras_dedup_key(kind, project.projcode, action_id,
-                                     recipient.address),
-            requested_by=current_user.username,
-        )
-        for recipient in to_recipients(people)
-    ]
+    return build_xras_messages(db.session, project, people, action=action,
+                               requested_by=current_user.username)
 
 
 @bp.route('/xras_notify_form/<int:project_id>')
@@ -1845,7 +1699,8 @@ def xras_notify_form(project_id: int):
     if project is None:
         return htmx_modal_not_found('Project')
 
-    action = _load_xras_action(request.args.get('action_id', type=int))
+    action = load_xras_action(db.session,
+                              request.args.get('action_id', type=int))
     people = get_xras_pending_recipients(db.session, [project_id]).get(project_id, [])
     messages = _xras_messages(project, people, action=action)
 
@@ -1938,7 +1793,8 @@ def xras_notify(project_id: int):
     if project is None:
         return htmx_not_found('Project')
 
-    action = _load_xras_action(request.args.get('action_id', type=int))
+    action = load_xras_action(db.session,
+                              request.args.get('action_id', type=int))
     people = get_xras_pending_recipients(db.session, [project_id]).get(project_id, [])
     messages = _xras_messages(project, people, action=action)
